Tests/LibWeb/Text/expected/Editing/select-all-editing-host-boundaries.txt

Atomic editing-host boundaries: anchor=div#first@0 focus=div#first@3 text="middle"
Script-moved selection: anchor=#text("other")@0 focus=#text("other")@5 text="other"
